Evan suggest to just overwrite the lite version by Stephane
